docker/requirements.txt
langchain-community==0.0.13 
langchain==0.1.1 
pypdf==3.17.4
sentence_transformers==2.2.2
transformers==4.39.3
langdetect==1.0.9
deep-translator==1.11.4
torch==2.1.2
gradio==4.25.0
diffusers==0.27.2
pydantic==2.6.4
qdrant_client==1.9.0
pillow==10.2.0
accelerate